backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.3
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
python-multipart>=0.0.6
